requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic>=2.3.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
joblib>=1.3.2